Fix for trac ticket [https://trac.osgeo.org/mapguide/ticket/88] - Native 64bit MapGuide Support http://trac.osgeo.org/mapguide/ticket/88
Changes:
- Fix size_t portability issue - on 64bit systems this is 64 bits in size and not 32 bits.

Notes:
- This is only a partial fix as there are still more issues to resolve.


git-svn-id: http://svn.osgeo.org/mapguide/trunk@3460 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ MgDev/Common/Foundation/Data/ByteSink.cpp
@@ -118,11 +118,11 @@ void MgByteSink::ToStream(ACE_SOCK_Stream* stream)
 {
     unsigned char bytes[4096];
 
-    int bytesReceived, bytesSent;
+    size_t bytesReceived, bytesSent;
 
-    while((bytesReceived = m_reader->Read(bytes, 4096)) > 0)
+    while((bytesReceived = (size_t)m_reader->Read(bytes, 4096)) > 0)
     {
-        if ((bytesSent = (int)stream->send_n(bytes, bytesReceived, MG_MSG_NOSIGNAL)) != bytesReceived)
+        if ((bytesSent = stream->send_n(bytes, bytesReceived, MG_MSG_NOSIGNAL)) != bytesReceived)
         {
             //should get the error from ACE and put it in the exception
             throw new MgStreamIoException(L"MgByteSink.ToStream",
@@ -223,11 +223,11 @@ void MgByteSink::ToFile(CREFSTRING filename)
 
         bytes = new unsigned char[BSINK_BUFFER_SIZE];
 
-        int bytesReceived, bytesWritten;
+        size_t bytesReceived, bytesWritten;
 
-        while ((bytesReceived = m_reader->Read(bytes, BSINK_BUFFER_SIZE)) > 0)
+        while ((bytesReceived = (size_t)m_reader->Read(bytes, BSINK_BUFFER_SIZE)) > 0)
         {
-            if((bytesWritten = (int)ACE_OS::fwrite(bytes, 1, bytesReceived, f)) != bytesReceived)
+            if((bytesWritten = ACE_OS::fwrite(bytes, 1, bytesReceived, f)) != bytesReceived)
             {
                 ACE_OS::fclose(f);
